+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Mistmatching token signing algorithms do not trigger service reg refresh [#88](https://github.com/rokwire/core-auth-library-go/issues/88)
+
 ## [3.0.0] - 2023-03-09
 ### Added
 - BREAKING: Encapsulate key management [#76](https://github.com/rokwire/core-auth-library-go/issues/76)